Safari-609.1.11.1/LayoutTests/platform/ios/imported/w3c/web-platform-tests/intersection-observer/cross-origin-iframe-expected.txt



PASS Intersection observer test with no explicit root and target in a cross-origin iframe. 
PASS First rAF 
PASS topDocument.scrollingElement.scrollTop = 200 
FAIL iframeDocument.scrollingElement.scrollTop = 250 assert_equals: expected 1 but got 0
FAIL topDocument.scrollingElement.scrollTop = 100 assert_equals: expected 1 but got 0
